Title:fix, restore failed when stop empty trash
Description:bug-30231

RootCause:Null

Solution:check delete result, rm info dir if ret is true

+++ dde-file-manager-lib/controllers/trashmanager.cpp
@@ -435,11 +435,13 @@ void TrashManager::cleanTrash(const QObject *sender) const
     if (QFile::exists(file_url.toLocalFile())) {
         list << file_url;
     }
-    fileService->deleteFiles(sender, list, false, false, true);
+    bool ret = fileService->deleteFiles(sender, list, false, false, true);
 
     // 清空回收站意味着回收站所有文件都被删除，因此直接删除info的目录即可
-    QString infoPaht = info_url.toLocalFile();
-    QProcess::execute("rm -r \"" + infoPaht.toUtf8() + "\"");
+    if (ret) {
+        QString infoPaht = info_url.toLocalFile();
+        QProcess::execute("rm -r \"" + infoPaht.toUtf8() + "\"");
+    }
 }
 
 bool TrashManager::isEmpty()
